requirements.txt
langchain
langchain_together
langchain_community
unstructured[all-docs]
python-dotenv
opencv-python
pyaudio
pillow
openai-whisper
faster_whisper
sox
SpeechRecognition
pygame
ane_transformers==0.1.3
coremltools
pyowm
exa_py
rich
streamlit
onnx==1.15